+++ packages/schema/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @minimajs/schema
 
+## 1.4.0
+
+### Minor Changes
+
+- 8874965: openapi support
+
+### Patch Changes
+
+- Updated dependencies [8874965]
+  - @minimajs/server@0.9.0
+
 ## 1.3.0
 
 ### Minor Changes